--- a/ggml/src/ggml-vulkan/vulkan-shaders/im2col.comp
+++ b/ggml/src/ggml-vulkan/vulkan-shaders/im2col.comp
@@ -68,8 +68,8 @@ void main() {
             continue;
         }
 
-        const int iiw = int(current_ix) * p.s0 + int(current_kx) * p.d0 - p.p0;
-        const int iih = oh_s1 + int(current_ky) * p.d1 - p.p1;
+        const uint iiw = int(current_ix) * p.s0 + int(current_kx) * p.d0 - p.p0;
+        const uint iih = oh_s1 + int(current_ky) * p.d1 - p.p1;
 
         offset_dst[idx] = dst_base + current_ix * p.CHW + current_ky * p.KW + current_kx;