@0cc4m 2025-02-13
If cswave32 helps you, maybe look into if setting the `requiredSubgroupSize` in the shader does the same thing.