@daniandtheweb 2025-02-13
I've been trying to use the VK_EXT_subgroup_size_control extension in GLSL to set the `requiredSubgroupSize` but I can't manage to make it work. Is the right approach to enable the "GL_EXT_subgroup_size_control" extension and then appending `requiredSubgroupSize` to the layout in which local_size_x_id is set or am I trying to make use the extension wrongly?