@0cc4m 2025-02-13
Oh sorry, I thought you knew that the extension is already implemented. You can set values when loading pipelines: