@daniandtheweb 2025-02-13
Thanks a lot for the information, I've been trying to implement it directly in the shader not knowing it was already implemented in the main ggml-vulkan.cpp code. 

I did some tests by manually setting the subgroup size for 32 and I can recreate the results I had using the RADV env.
I also tested creating a second pipeline which uses instead 64 as subgroup size and used it only on im2col and I managed to get another speedup on stable-diffusion (1.55 it/s PR + subgroup 32 vs 1.59 it/s PR + mixed subgroups). Do you think it could be a good idea creating two pipelines (1 with subgroup 64 and 1 with subgroup 32) and use them only on specific GPUs? I don't know if it could be useful on other GPUs.