@daniandtheweb 2025-02-22
Here are my findings:

1. 
```
    vk::PhysicalDeviceProperties2 props2;
    device->physical_device.getProperties2(&props2);
    std::string device_name = props2.properties.deviceName.data();
```
Declaring those three lines is needed to effectively enabling wave32, at least on RDNA1 (otherwise if I just call device->name directly everything just keeps using wave64).
2.
Setting `device->subgroup_size` as the same value as `required_subgroup_size` keeps wave32 results kinda neutral compared to wave64 (some exceptions are im2col, softmax). 
3.
Using the new changes I reach the same performance in llama-bench as with RADV_PERFTEST=cswave32 on master but without any additional flag.


`./llama-bench -m ~/Applications/chat/gguf/llama-2-7b.Q4_0.gguf -ngl 100`
```
Master:
ggml_vulkan: 0 = AMD Radeon RX 5700 XT (RADV NAVI10) (radv) | uma: 0 | fp16: 1 | warp size: 64 | shared memory: 65536 | matrix cores: none
| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 100 |         pp512 |        434.46 ± 0.67 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 100 |         tg128 |         68.75 ± 0.03 |

build: 51f311e0 (4753)
``` 
```
Master + cswave32:
ggml_vulkan: 0 = AMD Radeon RX 5700 XT (RADV NAVI10) (radv) | uma: 0 | fp16: 1 | warp size: 64 | shared memory: 65536 | matrix cores: none
| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 100 |         pp512 |        465.67 ± 0.33 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 100 |         tg128 |         67.72 ± 0.04 |

build: 51f311e0 (4753)
``` 
```
PR:
ggml_vulkan: 0 = AMD Radeon RX 5700 XT (RADV NAVI10) (radv) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 65536 | matrix cores: none
| model                          |       size |     params | backend    | ngl |          test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 100 |         pp512 |        465.67 ± 0.12 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | Vulkan     | 100 |         tg128 |         67.68 ± 0.26 |

build: 4a3988e9 (4688)
``` 


I have updated the branch wave_test in my stable-diffusion.cpp fork to use the new changes.

```
stable-diffusion.cpp sync : 
sd 1.5 - 512x512 - 20 steps - 16.52s, 1.41it/s
sd xl - 1024x1024 - 20 steps - 108.78s, 4.35s/it, vae tiling 2.65it/s
``` 
```
stable-diffusion.cpp wave_test:
sd 1.5 - 512x512 - 20 steps - 14.87s, 1.60it/s
sd xl - 1024x1024 - 20 steps - 101.46s, 4.02s/it, vae tiling 2.74it/s
```